Refactor Persistence layer in Infrastructure
Renamed Repositories and moved things into a Persistence folder to keep consistency
