+++ config/packages/doctrine_migrations.yaml
@@ -1,3 +1,3 @@
 doctrine_migrations:
-    dir_name: '%kernel.project_dir%/src/Infrastructure/Share/Migrations'
-    namespace: DoctrineMigrations
+    dir_name: '%kernel.project_dir%/src/Infrastructure/Share/Persistence/Doctrine/Migrations'
+    namespace: App\Infrastructure\Share\Persistence\Doctrine\Migrations
