+++ config/persistence/mapping/domain/User.ValueObject.Auth.Credentials.orm.xml
