+++ config/persistence/mapping/projections/User.Query.Projections.UserView.orm.xml
