+++ src/Application/Query/User/FindByEmail/FindByEmailHandler.php
@@ -7,15 +7,15 @@ namespace App\Application\Query\User\FindByEmail;
 use App\Application\Query\Item;
 use App\Application\Query\QueryHandlerInterface;
 use App\Domain\Shared\Query\Exception\NotFoundException;
-use App\Infrastructure\User\Query\Mysql\MysqlUserReadModelRepository;
+use App\Infrastructure\User\Query\Mysql\MysqlReadModelUserRepository;
 use App\Infrastructure\User\Query\Projections\UserView;
 use Doctrine\ORM\NonUniqueResultException;
 
 class FindByEmailHandler implements QueryHandlerInterface
 {
-    private MysqlUserReadModelRepository $repository;
+    private MysqlReadModelUserRepository $repository;
 
-    public function __construct(MysqlUserReadModelRepository $repository)
+    public function __construct(MysqlReadModelUserRepository $repository)
     {
         $this->repository = $repository;
     }
