+++ src/Infrastructure/Share/Persistence/Doctrine/Types/DateTimeType.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Infrastructure\Share\Doctrine;
+namespace App\Infrastructure\Share\Persistence\Doctrine\Types;
 
 use App\Domain\Shared\Exception\DateTimeException;
 use App\Domain\Shared\ValueObject\DateTime;
