+++ src/Infrastructure/User/Auth/AuthProvider.php
@@ -6,7 +6,7 @@ namespace App\Infrastructure\User\Auth;
 
 use App\Domain\Shared\Query\Exception\NotFoundException;
 use App\Domain\User\ValueObject\Email;
-use App\Infrastructure\User\Query\Mysql\MysqlUserReadModelRepository;
+use App\Infrastructure\User\Query\Mysql\MysqlReadModelUserRepository;
 use Assert\AssertionFailedException;
 use Doctrine\ORM\NonUniqueResultException;
 use Symfony\Component\Security\Core\User\UserInterface;
@@ -14,9 +14,9 @@ use Symfony\Component\Security\Core\User\UserProviderInterface;
 
 class AuthProvider implements UserProviderInterface
 {
-    private MysqlUserReadModelRepository $userReadModelRepository;
+    private MysqlReadModelUserRepository $userReadModelRepository;
 
-    public function __construct(MysqlUserReadModelRepository $userReadModelRepository)
+    public function __construct(MysqlReadModelUserRepository $userReadModelRepository)
     {
         $this->userReadModelRepository = $userReadModelRepository;
     }
